Bootstrap the template folder
when running the init command, you have the ability to select a custom template directory.
These changes, check if the directory exists, and contains files
If it doesn't exist it will ask if it should be created, then it will ask if you want to populate files
If you say yes, the default templates packed with the tool, will be populated into your custom template directory so you can get straight to editing them. have fun!