More work for #37
Spectre console for command apps is great but as requirements evolve, so does the need to have more control. The pretty much defacto standard for writing cli apps is System.CommandLine, so i've switched to using that, with extra middleware to give a 'truer' dependency injection experience.
Still using AnsiConsole from spectre console, as I think its awesome, for for option management and commands, we're now using System.CommandLine