This introduces non interactive support for the init command, and the base for others.
Only implemented for the init command for now
We retroactively apply any passed in args now into state, pre execution of a command.
There is a new BaseAction type that adds non interaction validaton
Then its a case of accessing NOnInteractive from State, and handling in the action. Should be easy to apply to other actions now.