Refactor global usings and adjust assembly info
The global usings in `GlobalUsings.cs` have been refactored to enhance the organization of the code. Some of the usings have been removed or replaced for better maintenance and readability. Additionally, the assembly and tool name in `Aspirate.Cli.csproj` have been changed to lowercase to ensure consistency across the codebase.

Handles issue reported in #65