Refactor manifest actions and improve secret handling
Code was refactored to improve the handling of manifests for clusters with the introduction of handling for secret files. Secret files are now created, used and cleaned up appropriately when interacting with a Kubernetes cluster. Additionally, a small tweak was made to the selection prompts used in various actions, adding a blue highlight style for better user experience.
This concludes the final part of secret integration for initial secret support. Going to require some feedback this one...