Inputs and Secrets (#66)
* Added container processor  -  will work on inputs when i feel better

Move where manifest value substitution takes place - should be responsibility of the processor - not the file loader.

Ensure all services are documented

connectionstring parsing

Create SubstituteValuesAspireManifestAction for placeholder substitution

Placeholder substitution in Aspire manifests is now handled by a new, dedicated action called SubstituteValuesAspireManifestAction. This change provides an improvement in flexibility and extensibility of the substitution process. Container inputs are now supported and corresponding unit tests have been updated.

Update connection string handling in processors

Revamped the handling of connection strings in Aspirate's processing logic. ConnectionString placeholders are now replaced with actual values in the `BaseProcessor` to account for different types of resources, such as containers. Also updated relevant unit tests to reflect these changes.

* Refactor global usings and adjust assembly info

The global usings in `GlobalUsings.cs` have been refactored to enhance the organization of the code. Some of the usings have been removed or replaced for better maintenance and readability. Additionally, the assembly and tool name in `Aspirate.Cli.csproj` have been changed to lowercase to ensure consistency across the codebase.

Handles issue reported in #65

* Add password generation and prompt for non-generated secrets

This update enhances Aspirate's security capabilities by integrating a password generator and providing a prompt for non-generated secrets. In addition, several objects have been extended with new interfaces for handling inputs and bindings, and appropriate adjustments have been made in code to accommodate these changes.

This is the first pass phase of secret support. Right now inputs are generated, and still supplied as Environmental variables. All ConnectionStrings will be moved into generated opaque secrets as part of phase 2.

Phase 3 will introduce state management of the secrets storing them as Aes encrypted values.

* Improve user input handling for secrets

The process of handling manually entered and generated inputs has been refined to enhance the user experience. The inputs now include a double-confirmation to avoid mismatches, and a clear success message when a value is successfully generated. The `IResourceWithInput` interface was also updated to include the resource's Name.

* Add PopulateInputs and SaveInputs actions

These changes replace the PromptForNonGeneratedSecrets action with two separate actions: PopulateInputs and SaveInputs. The PopulateInputs action handles both manually entered and automatically generated secrets, while SaveInputs was added as a placeholder for a yet-to-be-defined functionality. This separation enhances code organization and readability. The command flow in GenerateCommandHandler is also updated to reflect this change. Tests for these new actions are also included.

* Implement multi-level secret structure and enhance secret provider

This update implements a two-level secret structure in Commands and Secrets modules to support separate resources for managing secrets. A resource feature has been added to the SecretProvider and tests were updated to reflect these changes. This allows each individual resources secrets to be stored separate of each other - to allow for the same input name across containers.

The secret provider has also been changed to accept a file path rather than the raw state for secret state loading.

* Add SecretStateExists method and tests

Implemented a new method SecretStateExists to check the existence of a secret state in the ISecretProvider interface. This enables detecting if a secret state file exists in the current directory. Also, added corresponding unit tests to verify the functionality for both PasswordSecretProvider and Base64SecretProvider.

* Add SaveSecretsAction and related tests

This commit introduces the SaveSecretsAction along with its respective unit tests. This action includes creating and verifying passwords, asking users for existing secrets or overwriting them, and saving the state of secrets. The tests include different scenarios to ensure correct and robust functioning of this action.

* Update secrets handling in SaveSecretsAction

The commit introduces checks for existence of resources and secrets before performing operations. This optimizes secret saving process by preventing unnecessary purging or reflexivity operations. It also includes more extensive testing, ensuring correctness of the process.
Also in this commit - I've changed what actually represents a secret. In my previous iteration, i treated inputs directly as secrets - but these always exist in the manifest - what we should be treating as secrets are the transposed connection strings.

* Add LoadSecretsAction and accompanying tests

Implemented a new action, LoadSecretsAction, within Aspirate.Commands to handle the loading of secrets and password validation. The new implementation is accompanied by a series of unit tests to validate the correct operation of the added functionality.

* Add secret password option to Apply commands

This commit incorporates a secret password option into the "Apply" commands. Relevant changes include the introduction of the option in SharedOptions and ApplyOptions. Additionally, relevant tests have been updated to accommodate this change.

* Add new ConnectionStringProtector and improve secret handling

Wasn't happy with the way SaveSecretsAction turned out. It felt clunky to use in the app over relaunched instances.

Cleaned up the implementation.

Added the strategy pattern with just a ConnectionStringProtector class implementation for now, for managing connection string secrets. Streamlined secret protection workflow in SaveSecretsAction using a collection of protection strategies, and improved password check system with more clear terminal prompts. Added functionality to delete existing secrets.
Updated secrets.

Feels better.

* Refactor manifest actions and improve secret handling

Code was refactored to improve the handling of manifests for clusters with the introduction of handling for secret files. Secret files are now created, used and cleaned up appropriately when interacting with a Kubernetes cluster. Additionally, a small tweak was made to the selection prompts used in various actions, adding a blue highlight style for better user experience.
This concludes the final part of secret integration for initial secret support. Going to require some feedback this one...