Implement secret state check and optional registry for Dockerfiles
The commit adds a check for secret state before processing secretProvider in the RemoveManifestsFromClusterAction and ApplyManifestsToClusterAction classes. Also, it makes the Docker container registry optional across different classes, appending the registry to the image tag only when provided. Finally, an imagePullPolicy was included in Dockerfile processors to manage Docker image pulling more effectively.
This fixes #70