83 Deploying eshop fails with naming problem (#86)
* Normalize casing for container image and name fields

The container image and name fields in the ProjectTemplateData, DockerfileTemplateData, and BaseTemplateData classes have been normalized to lowercase. This change ensures consistent casing across the application and helps prevent potential casing-related issues in the future.

* Add namespace support for Kubernetes deployments

Added code to allow generating a Kubernetes Namespace resource and applying it to all generated resources. This namespace will be used at deployment time. This feature allows users to specify a namespace during the generation process which adds more flexibility and organization to Kubernetes deployments.