Add builder parameter to BuildAndPushContainerForProject
The BuildAndPushContainerForProject method in ProjectProcessor and ContainerCompositionService have been updated to include a new builder parameter, allowing to specify 'docker' or 'podman'. In addition, tests have been added to ensure that an exception is thrown when the builder is offline. Also, checks have been implemented in ShellExecutionService to deal with situations when the builder is not running.