Add Dapr processors and related Literals
This commit introduces new processors for handling Dapr components and systems in the Aspirate.Processors project. A corresponding literals are added in AspireComponentLiterals. It also includes new models for Dapr manifests in AspireManifests. No longer needed 'Resources' folder reference is removed from the project file.

Rename resources, and move into structured groupings - common types in folders.

Replace ProtectableLiterals with ProtectorType using SmartEnum

Ardalis.SmartEnum.SystemTextJson package was added and ProtectableLiterals class was replaced with SmartEnum implementation, ProtectorType. As a result, protectors and processors were refactored to use the new enumeration instead of the ProtectableLiterals class that was deleted. The change improves type-safety and readability of source code in the project.

Add support for applying Dapr annotations

The updates allow Dapr annotations to be applied to services in the Aspirate processing system. A new action, ApplyDaprAnnotationsAction, has been added to handle applying annotations, and the capability to handle/test for these annotations has been added to the appropriate sections of the code. This enhancement should improve the interoperability with the Dapr system.

Add Dapr annotations to Aspirate tests

Expanded the GlobalUsings.cs module and added Dapr annotations to the file's manifest within the Aspirate test suite. Reflecting these changes, more tests have also been added into the ManifestFileParserServiceTests.cs

Add Dapr component to manifest file and improve code organization

This update adds support for Dapr components in the manifest file. It also simplifies the container creation task by reformatting parameters for better readability. Unused code imports across various files are removed to maintain a cleaner codebase. Changes also include general codebase improvements like making some methods more organized and grouping similar functionalities together.

Add Dapr CLI service and improve handling of Dapr components

This commit introduces a Dapr CLI service in Aspirate.Services with methods to check if Dapr is installed and to install it. It also improves Dapr component handling by adjusting the template and including Dapr files in the final manifests. Moreover, a safer handling of Postgres environment password in PlaceholderSubstitutionStrategies is implemented.

Add functionality to remove Dapr from cluster

This commit introduces functionality to check for Dapr during the removal of manifests from the cluster. If Dapr is found in the cluster, the user will be asked whether they would like to remove it. If the user chooses to remove Dapr and the operation fails, an error message will be logged and the action will exit. This update enhances the clarity and effectiveness of the cluster clean-up process.

Update launchSettings.json command arguments and directories

Modified the commandLineArgs and workingDirectory paths for the generate, apply, and destroy keys in launchSettings.json. The commandLineArgs now only contains "generate", and the directories are updated to the eShopLite project directory.

Add check for Dapr CLI installation before operations

A check for Dapr CLI installation has been added before manifest removal or application to the cluster. If it's not installed, relevant informative messages are displayed to the user. Additionally, a method to check for command availability has been implemented in `IShellExecutionService` and `ShellExecutionService`.