Add Docker compose file generation functionality
This update includes the ability to generate a Docker compose file based on the supported components selected. It adds a new method to process each component individually and write the output to a file. The output path for this file has been made a constant for easy reference. Relevant dependency to Aspirated.DockerCompose project has also been included. Changes are made to generate the compose entries in CreateComposeEntry method and necessary references have been included into globalUsings.cs file. This new feature improves the automation process and increases the ease of deployment.