127 `aspirate apply` doesn't respect the `--image-pull-policy Always` option from `aspirate generate` (#132)
* Add rolling restart functionality to KubeCtlService

Implemented a method in KubeCtlService for performing a rolling restart in a Kubernetes environment. Also updated related interfaces, command options, and added a toggle switch in Apply command. Included test cases for the new feature.

* Update deployment strategy and restart options

The deployment strategy has been changed from 'rollingUpdate' to 'Recreate' in the deployment templates. The 'RollingRestart' option in Apply is now set to default of false, and is purely opt-in by passing -r.