Remove Azure processors and update ports mapping (#149)
This update removes the AzureKeyVaultProcessor, AzureStorageBlobProcessor, and AzureStorageProcessor, as they are no longer needed. In addition, it revises the handling of ports in ContainerProcessor, DockerfileProcessor, and ProjectProcessor by introducing new extension methods. This change makes the code cleaner and more manageable, improving readability and maintainability. Furthermore, a test for docker file processing using nodejs.json is added, and slight imageName parsing changes have been added to DockerfileProcessor.cs to ensure that the correct imageName exists in the image cache