Refactor KubernetesDeploymentTemplateData and reshape related operations
We've restructured the KubernetesDeploymentTemplateData into a more appropriate place and renamed it to KubernetesDeploymentData. Notable changes include its relocation from the Aspirate.Processors namespace to Aspirate.Shared.Models.Aspirate, the addition of some processing methods relevant to Kubernetes deployment data, and cleaning up the dependencies in the files that use this class. Subsequent actions such as creating Kubernetes objects, validating data and setting the properties of KubernetesDeploymentData can now be performed in a more organized and clear manner.

We've also introduced extension based mapping from the KubernetesDeploymentData to raw Kubernetes models utilising the c# KubernetesClient