Refactor code and update state properties
Deleted unnecessary files in the Azure directory under Models folder. Removed unused using directives throughout the code. Added a new attribute `RestorableStateProperty` to identify properties in the state that can be persisted and restored. Scaling down redundant dependencies helps improve the maintainability and readability of the code, and enhance the performance.

State is now persisted across runs. More work required still to work this
into actions, and a choice to be made whether to store this as as plain text json, or write it as a byteArray?
 Will sleep on it..