Aspirate State Management (#175)
* Refactored Enums and related classes to Shared project

The Enums classes and related usings along with Processor-related interfaces have been moved to Shared project from Commands. Unused usings for Secrets.Literals have been removed from throughout the codebase

* Added Kubernetes object creation ability to resource processor

Created a new method, CreateKubernetesObjects, in the IResourceProcessor interface, and added a default implementation in the BaseResourceProcessor class. This allows for creating Kubernetes objects in the Aspirate application. Also, defined a new CreateKubernetesObjectsOptions class which encapsulates the options that are used by the CreateKubernetesObjects method.

* Refactor KubernetesDeploymentTemplateData and reshape related operations

We've restructured the KubernetesDeploymentTemplateData into a more appropriate place and renamed it to KubernetesDeploymentData. Notable changes include its relocation from the Aspirate.Processors namespace to Aspirate.Shared.Models.Aspirate, the addition of some processing methods relevant to Kubernetes deployment data, and cleaning up the dependencies in the files that use this class. Subsequent actions such as creating Kubernetes objects, validating data and setting the properties of KubernetesDeploymentData can now be performed in a more organized and clear manner.

We've also introduced extension based mapping from the KubernetesDeploymentData to raw Kubernetes models utilising the c# KubernetesClient

* Refactor Kubernetes objects creation into one method

The repetitive logic responsible for creating various Kubernetes objects such as ConfigMap, Secret, StatefulSet, Deployment, and Service was extracted into a single method called `ToKubernetesObjects()`. This method is part of the `KubernetesDeploymentData` class and unifies the creation of these objects, enhancing the readability and maintainability of the code.

* Add KubernetesDeploymentDataExtension tests

A new test suite, KubernetesDeploymentDataExtensionTests, has been added which includes multiple unit tests validating the functionality of the KubernetesDeploymentDataExtension class. Also, the k8s.Models namespace is globally used within Aspirate.Tests for simplicity.

* Extend KubernetesDeploymentDataExtensionTests functionality

Added support for configuration maps and secrets to the KubernetesDeploymentDataExtensionTests. The extension now verifies the shared configuration and secrets along with other existing Kubernetes objects like V1Deployment and V1Service.

* Update Kubernetes StatefulSet test to include volume checks

The unit test for the Kubernetes stateful set in the KubernetesDeploymentDataExtensionTests has been updated to include checks for volumes. Previously, the test was only checking name and image properties; now it verifies the volume properties as well.

* Add mount path check to Kubernetes extension tests

In the Kubernetes deployment data extension tests, a check was added to verify that the mount path of the first volume mount in the first container is correctly set to "/data". This ensures that the mount path is being correctly set in the system under test.

* Refactor code and update state properties

Deleted unnecessary files in the Azure directory under Models folder. Removed unused using directives throughout the code. Added a new attribute `RestorableStateProperty` to identify properties in the state that can be persisted and restored. Scaling down redundant dependencies helps improve the maintainability and readability of the code, and enhance the performance.

State is now persisted across runs. More work required still to work this
into actions, and a choice to be made whether to store this as as plain text json, or write it as a byteArray?
 Will sleep on it..

* Refactor state management in Aspirate

This commit removes the LoadStateAction and SaveStateAction functionality from the Aspirate.Commands section and integrates it directly into the command execution pipeline, ensuring the saved state is utilized and updated every time a command is run. It also includes the necessary changes to the relevant classes and methods to support this refactor, alongside a rename in AspirateLiterals for the state file name. This ensures that state is restored prior to options passed in to a cli run session being applied to state. State is restored, then the state is mutated based on passed in parameters.

* Refactor code to improve secrets management and update paths

This commit refactors some code to improve the management of secrets. It fixes issues with incorrect paths by updating the references to the new paths in related files. Also, some unused dependencies were removed from several files contributing to cleaner code. Unit tests were therefore updated to reflect these changes.

The next change that will be implemented will be the movement of the secrets from aspirate-secrets.json in the output path, into the aspirate-state.json file in the AppHost project directory.

* Remove base64 and password-based secret providers

This commit eliminates the implementation of specifically base64 and password-based secret providers from the codebase. The changes include the removal of related classes and tests. The remaining "SecretProvider" class is made more generic to handle secrets without specific encoding or encryption methods.

base64 secret provider was about as useful as a chocolate fire guard. I mean, it was purely an unencrypted string.

* Update SecretProvider and related functionality

The SecretProvider class and related calls have been updated. Now, instead of dealing with file paths for the secret state, the SecretProvider uses AspirateState instances directly. This change simplifies state handling and improves code readability. Some less used public properties in SecretProvider were also removed to streamline the class interface. Code related to privateRegistryPassword has also been adjusted.

* Update Aspirate module for enhanced command line interaction

This commit includes updates to the Aspirate module to enhance its command line interaction. The primary changes involve adding and adjusting options for various operations such as build, skip, and manifest generation. Additionally, unnecessary dependencies are removed, and non-required conditions are simplified. Some test adjustments are made to reflect these changes.

This wraps up addition of state management, which is now additive based on previous run - augmented by newly passed in commandline option, and state is shared across all aspirate commands (except init)