Update nullability and inference logic for environment variables (#178)
The code has been updated to allow null values in environment variables and secrets dictionaries. It also refactors retrieval methods for both environmental and secret variables. Now, instead of directly calling methods on object instances, it utilizes a pair of the resource key and object as function parameters for greater flexibility. The dashboard injection logic has been updated to avoid duplication. The unit tests have been adapted to reflect these changes.
Otlp service name is now included and set to the resource key from the manifest (same for docker compose) which should fix the display on the dashboard.