Update ContainerCompositionService and DotNetSdkLiterals (#220)
This commit introduces changes to the ContainerCompositionService and DotNetSdkLiterals source files. Specifically in the ContainerCompositionService file, it modifies the PublishProfileArgument to ContainerTargetArgument, with an empty string value. A few additions were also made to the DotNetSdkLiterals file, including the addition of the ContainerTargetArgument and the removal of the ContainerPublishProfile constant.